web/requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.10.4
gunicorn==19.3.0
itsdangerous==0.24
MySQL-python==1.2.5
Flask-MySQL==1.3.0
httplib2==0.9.2
Flask-SQLAlchemy==2.0.0
apscheduler